refactor(feishu): 历史拉取群配置移到消息监控台「历史消息处理」下方
历史拉取群属于「群历史」语义，放消息监控台的历史消息处理卡片下方更贴切，
集中到一个配置入口。从智能体配置抽屉（AssistantConfigDrawer）与详情页
（AssistantDetailPage）撤掉，只留消息监控台（WorkspaceSettingsPanel）一处，
按当前工作空间的飞书 bot 配置。
